+++ packages/orq-rc/src/lib/config.ts
@@ -54,8 +54,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "1.43.0-rc.1",
-  genVersion: "2.461.4",
+  sdkVersion: "1.43.1-rc.0",
+  genVersion: "2.462.0",
   userAgent:
-    "speakeasy-sdk/typescript 1.43.0-rc.1 2.461.4 2.0 orq-poc-typescript-multi-env-version",
+    "speakeasy-sdk/typescript 1.43.1-rc.0 2.462.0 2.0 orq-poc-typescript-multi-env-version",
 } as const;