+++ packages/orq-rc/src/models/errors/index.ts
@@ -3,6 +3,7 @@
  */
 
 export * from "./apierror.js";
+export * from "./deletev2resourcesevaluatorsid.js";
 export * from "./honoapierror.js";
 export * from "./httpclienterrors.js";
 export * from "./sdkvalidationerror.js";